+++ Modix.Services/Promotions/DBPromotionRepository.cs
@@ -19,12 +19,12 @@ namespace Modix.Services.Promotions
 
         public async Task AddCampaign(PromotionCampaign campaign, SocketGuildUser user)
         {
-            var promoUser = await _context.DiscordUsers.FirstOrDefaultAsync(u => (ulong) u.DiscordUserId == user.Id);
+            var promoUser = await _context.DiscordUsers.FirstOrDefaultAsync(u => u.DiscordUserId == user.Id);
             if (promoUser == null)
                 await _context.DiscordUsers.AddAsync(new DiscordUser
                 {
                     Username = $"{user.Username}#{user.Discriminator}",
-                    DiscordUserId = (long) user.Id,
+                    DiscordUserId = user.Id,
                     CreatedAt = DateTime.UtcNow,
                     IsBot = false,
                     Nickname = user.Nickname,
